edit

+++ src/Data/Array/Accelerate/TypeLits/Internal.hs
@@ -163,8 +163,7 @@ mean' ::
   => Acc (Array sh e)
   -> AccScalar c
 mean' x =
-  AccScalar $
-  A.unit $ realToFrac (A.the (A.sum (A.flatten x))) / realToFrac (A.size x)
+  AccScalar $ A.unit $ A.the (A.sum (A.flatten x)) / fromIntegral (A.size x)
 
 mkVector ::
      forall n a. (KnownNat n, Elt a)
